writeup 695 (CONTROL PHASE #2: two knobs dominantly separable + sign-stable but NOT strictly independent -- multiplicative interaction on the softmax; g's entropy effect grows 1.57->4.06 with alpha; null held, cross-axis 12x asymmetric) + fix rspd_mlp17_core_readout dtype (wf double->float) and requeue

+++ basis_aligned/bilinear_quotient/rspd_mlp17_core_readout.py
@@ -62,7 +62,7 @@ def capture_gate_and_wfreq(fresh):
     h1.remove(); h2.remove()
     Xg = torch.cat(cap, 0); Oc = torch.cat(O, 0); lfc = torch.cat(lf, 0)
     wf = ((Oc - Oc.mean(0)) * (lfc - lfc.mean())[:, None]).mean(0)
-    wf = wf / wf.norm()
+    wf = (wf / wf.norm()).float()
     return Xg, wf
